Refactor: path for images
