+++ src/assets/image/icon/index.js
@@ -1,0 +1,19 @@
+import notFound from "./notFound.png";
+import congratulations from "./congratulations.png";
+import desktop from "./desktop.png";
+import error from "./error.png";
+import heart from "./heart.png";
+import neural from "./neural.png";
+import signLanguage from "./signLanguage.png";
+
+const icon = {
+  notFound,
+  congratulations,
+  desktop,
+  error,
+  heart,
+  neural,
+  signLanguage,
+};
+
+export default icon;
