+++ src/assets/index.js
@@ -2,12 +2,14 @@ import list from "./image/list";
 import consonants from "./image/consonants";
 import vowels from "./image/vowels";
 import words from "./image/words";
+import icon from "./image/icon";
 
 const IMAGE = {
   list,
   consonants,
   vowels,
   words,
+  icon,
 };
 
 export default IMAGE;
